Change ProgrammaticModelStream to allow different data types

+++ mpp/src/SphereModelStream.cpp
@@ -101,7 +101,7 @@ namespace mpp
 		addTriangle(11, 2, 7);
 
 		// Resize data here, to stop this happening in subdivide
-		size_t scaleFactor = pow(4, (res - 1));
+		size_t scaleFactor = (size_t)pow(4, (res - 1));
 		//mMeshDataDefinition.vertexData.resize(mMeshDataDefinition.vertexData.size() * scaleFactor);
 		//mMeshDataDefinition.indexData.resize(mMeshDataDefinition.indexData.size() * scaleFactor);
 		for (int i = 1; i < res; ++i)
